IOS, Android 통합 리더보드 관리를 위해 GameServiceManage를 통한 연결방식으로 변경. 다운힐 결과 점수 리더보드에 반영

+++ Assets/Scripts/Manager/GameManager.cs
@@ -23,7 +23,7 @@ public class GameManager : Singleton<GameManager> {
 
     public float panelty_time;              //폴 통과하지 못한 경우 패널티
     public int startTime;
-    public int[] highestScores;                    //각 종목별 최고 점수
+    public float[] highestScores;                    //각 종목별 최고 점수
     
     private void Awake() {
         DontDestroyOnLoad(gameObject);
